[change] New spam related feature : the robot parameter "reporting_spam_script_path" is the path an external script dedicated to signal undetected spam. If set, list editor can tag message in moderation spool. Then they are piped to this external script (see contribs repository to have arf and imap examples).
In addition the list moderation web interface is renewed, the viewmod action is now an ajax feature in the modindex page. javascript is used to make display more simple and contextualized.

Ajax is introduced in wwsympa;fcgi (internal is similar to rss feature. grepa ajax and modindex and viewmod action are good example. 




git-svn-id: https://subversion.renater.fr/sympa/trunk@6796 05aa8bb8-cd2b-0410-b1d7-8918dfa770ce