[feature][#6089]There is now, four new params in a config list :Starthour,Startminute,Endhour, Endminute. There are implements in SQLSources, LDAPSources and LDPA2levelSources. With this params, you can enter a period of time wich disable the syncronization.
There are use in a new function, in "Datasoutce.pm" : "is_allowed_to_sync", which return 1 if the
 syncronization is allowed, or 0 if it is not. With this function, you can start your period of n
o syncronization the current day, and finish it, the next day.

Then, this function is called in the function "_load_list_members_from_include" in "List.pm"
Firstly, we check if the current id was present in the hash of old_subscribers* (it corresponds a
t the subscribers before the syncronization).
This permit to know if it's a new subscribers, OR if there is changing in the configuration list.
Next, we check if it's allowed to syncronize or not :
If it is : we can add new subscribers and update the others.
It it is not : we make an array with subscribers who are excluded.

To finish, in sync_include, we check if there are exclusions. 
If it is the case :
Feed the new_subscribers hash with users previously subscribed with data sources not used because
 we were not in the period of
time during which synchronization is allowed. This will prevent these users from being unsubscrib
ed.

*The hash of old_subscribers (wich is returned)is create in a new function "get_list_of_sources_i
d".
This function is called in parameters in the function "_load_list_members_from_include".



git-svn-id: https://subversion.renater.fr/sympa/branches/stage-yohann@7066 05aa8bb8-cd2b-0410-b1d7-8918dfa770ce